Add user profile and video owner

+++ src/controllers/videoController.js
@@ -1,5 +1,6 @@
 import { Model } from "mongoose";
 import Video from "../models/Video";
+import User from "../models/User";
 
 //promise
 export const home = async (req, res) => {
@@ -10,32 +11,43 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
   //const id = req.params.id;
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video) {
-    ////존재하지 않는 비디오 페이지로 접근할 경우 404 페이지 출력
     return res.status(404).render("404", { pageTitle: `Video not found.` });
   }
   return res.render("watch", { pageTitle: video.title, video });
 };
 
-//painting the Form
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: `Video not found.` });
   }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   return res.render("videos/edit", { pageTitle: `Edit ${video.title}`, video });
 };
+
 //Saving the Changes
 //video는 데이터베이스에서 검색한 영상 오브젝트
 //Video는 비디오모델
 export const postEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(404).render("404", { pageTitle: `Video not found.` });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -50,15 +62,22 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
       fileUrl,
       hashtags: Video.formatHashtags(hashtags),
+      owner: _id,
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("videos/upload", {
@@ -70,6 +89,12 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
